+++ tool_statistics/presets.py
@@ -8,6 +8,7 @@ from .style import (
     STYLE_VECTOR,
     STYLE_TIMESTEP_REDUCTION_ANALYSIS,
     STYLE_BALANCE,
+    STYLE_WATER_ON_STREET_DURATION_NODE,
 )
 
 
@@ -245,6 +246,25 @@ SOURCE_SINK_MM_PRESETS = Preset(
     raster_layer_name="Source or sink (raster)",
 )
 
+# Change in water level
+water_on_street_aggregations = [
+    Aggregation(
+        variable=AGGREGATION_VARIABLES.get_by_short_name("s1"),
+        method=AGGREGATION_METHODS.get_by_short_name("time_above_threshold"),
+    ),
+]
+
+WATER_ON_STREET_DURATION_PRESET = Preset(
+    name="Water on street duration",
+    description="Duration of water level above manhole drain level.",
+    aggregations=water_on_street_aggregations,
+    nodes_style=STYLE_WATER_ON_STREET_DURATION_NODE,
+    nodes_style_param_values={"column": "s1_time_above_threshold"},
+    nodes_layer_name="Water on street duration (nodes)",
+    raster_layer_name="Water on street duration (raster)",
+    only_manholes=True,
+)
+
 PRESETS = [
     NO_PRESET,
     MAX_WL_PRESETS,
@@ -252,4 +272,5 @@ PRESETS = [
     SOURCE_SINK_MM_PRESETS,
     FLOW_PATTERN_PRESETS,
     TS_REDUCTION_ANALYSIS_PRESETS,
+    WATER_ON_STREET_DURATION_PRESET,
 ]